Master branch updates for ESAPI-2.2.0.0-RC2

+++ src/main/java/org/owasp/esapi/waf/internal/InterceptingServletOutputStream.java
@@ -127,7 +127,7 @@ public class InterceptingServletOutputStream extends ServletOutputStream {
 						amountToWrite = (int) (totalSize - currentPos);
 					}
 			
-					out.read(buff, 0, (int)amountToWrite);
+					out.read(buff, 0, amountToWrite);
 					
 					os.write(buff,0,amountToWrite);
